move all component file
