+++ src/components/Cardticle.jsx
@@ -5,36 +5,33 @@ import { Col, Container, Row, Button } from "react-bootstrap";
 import "./../styles/article-home.css";
 
 function Cardticle() {
-  const [cards, setCard] = useState([])
+  const [cards, setCard] = useState([]);
 
   useEffect(() => {
-    axios(
-      "https://6480432af061e6ec4d48ebcc.mockapi.io/article-home"
-    ).then(result => setCard(result.data))
-  }, [])
+    axios("https://6480432af061e6ec4d48ebcc.mockapi.io/article-home").then((result) => setCard(result.data));
+  }, []);
 
   return (
     <>
-            {cards.map(item => (
-                <Col lg={{ span: 4, offset: 0 }} key={item.id} id="col-card">
-                {/* </Col><Col md={4} key={item.id} id="col-card"> */}
-                  <div>
-                    <Card className="card" id="card">
-                      <Card.Img variant="top" src={item.img} id="card-img"/>
-                      <Card.Body className="card-body">
-                        <Card.Title className="card-title">{item.title}</Card.Title>
-                        <Card.Text className="card-text">{item.desc}</Card.Text>
-                        <a href={item.link} id="card-goto-read">
-                          Baca Selengkapnya
-                        </a>
-                      </Card.Body>
-                    </Card>
-                  </div>
-                </Col>
-            ))}
-
+      {cards.map((item) => (
+        <Col lg={{ span: 4, offset: 0 }} key={item.id} id="col-card">
+          {/* </Col><Col md={4} key={item.id} id="col-card"> */}
+          <div>
+            <Card className="card" id="card">
+              <Card.Img variant="top" src={item.img} id="card-img" />
+              <Card.Body className="card-body">
+                <Card.Title className="card-title">{item.title}</Card.Title>
+                <Card.Text className="card-text">{item.desc}</Card.Text>
+                <a href={item.link} id="card-goto-read">
+                  Baca Selengkapnya
+                </a>
+              </Card.Body>
+            </Card>
+          </div>
+        </Col>
+      ))}
     </>
-  )
+  );
 }
 
 export default Cardticle;
