+++ src/components/CategoryDoctor.jsx
@@ -1,24 +1,13 @@
 import { Col, Container, Row, Button } from "react-bootstrap";
-import CardCategory from "../components/CardCategory";
+import CardCategory from "./CardCategory";
 import { Link } from "react-router-dom";
 import "./../styles/category-doctor.css";
 
 const CategoryDoctor = () => {
-
-  // const [category, setCategory] = useState([])
-
-  // useEffect(() => {
-  //   const getCategory = async () => {
-  //     const response = await fetch("http://localhost:3000/category");
-  //     const data = await response.json();
-  //     setCategory(data);
-  //   };
-
-
   return (
     <>
       <section id="card" className="category-card">
-        <div className="container" >
+        <div className="container">
           <div id="card-row" className="row justify-content-center">
             <CardCategory />
           </div>
@@ -27,7 +16,5 @@ const CategoryDoctor = () => {
     </>
   );
 };
-
-
 
 export default CategoryDoctor;
