--- src/pages/Landing.jsx
@@ -1,9 +1,0 @@
-const Landing = () => {
-    return (
-        <div>
-        <h1>Landing Page</h1>
-        </div>
-    );
-    }
-
-export default Landing;
